Backend/requirements.txt
alembic
bidict
certifi
cffi
charset-normalizer
click
colorama
dnspython
eventlet
Flask
Flask-Cors
Flask-JWT-Extended
Flask-Migrate
Flask-SocketIO
Flask-SQLAlchemy
gevent
gevent-websocket
greenlet
idna
importlib-metadata
importlib-resources
itsdangerous
jinja2
Mako
MarkupSafe
marshmallow
mysql-connector-python
protobuf
pycparser
PyJWT
python-dotenv
python-engineio
python-socketio
requests
six
SQLAlchemy
typing-extensions
urllib3
Werkzeug
zipp
zope.event
zope.interface